Label watershed pixels at push if compactness=0
See discussion at https://github.com/scikit-image/scikit-image/issues/2636

"""
When compactness is not a factor, the first version [labeling at push
time] is correct, because the labelling is determined solely by which
label gets there first. (If they get there at the same time, then it's
arbitrary and you might as well give it to the label that pushes first.)
But when compactness is a factor, labelling at push-time is a bug,
because a different neighbor could get there "later" but actually be
"closer" when compactness is factored in.
"""