Make `regionprops._slice` a user-facing property
Currently, `_slice` is hidden from the user, even though it would be
a useful property in many instances (for example, extracting objects
from one channel based on the regionprops of another). In fact there
was a discussion about this when we were talking about the bounding box
property, but that went nowhere.