Simplify the Dask Array test of `apply_parallel`
Now that a Dask Array will be preserved by `apply_parallel` (unless a
user specifically overrides it), update the test to drop the `compute`
argument in this case. As there already is a test to ensure the result
is a Dask Array, this should work as a test of the current behavior (in
addition to the NumPy case where we set `compute`).