Clarify that urllib objects are being mocked
Named importation forces us to mock the local name

skimage/io/util.py
from urllib.request import urlopen

skimage/io/tests/test_io.py
    ...
    monkeypatch.setattr('skimage.io.util.urlopen', mockopen)

we want to be clear that the mock is of an external library. this patch
achieves that at the expense not being as concise. Fair tradeoff