Refactor/move neighborhood utility functions in morphology (#4209)
* Move shared nd-utility functions into _util module

* Simplify/tweak helper functions in morphology._util

Simplify construction of indices, use consistent name for structuring
element (selem), some reformatting and improve documentation.

* Support raveled offsets for dimensions smaller selem

The old implementation would raise a ValueError if a dimension length
given in `image_shape` was smaller than in `selem`. This precluded
operations on images e.g. with the shape (2, 100).

* Address unintended reorering by np.unique

* Check dimensions _offsets_to_raveled_neighbors

* Enhance docstrings and examples

which may act as lightweight tests.

* Add test for _offsets_to_raveled_neighbors

* Add explicit _offsets_to_raveled_neighbors tests

The explicit tests are meant to catch subtle changes in the order of
neigbors with the same connectivity. They were reviewed by constructing
an image of the given shape and filling in the raveled image around the
image center and manual inspection of the result. E.g. with

    image = np.zeros(image_shape)
    image_raveled = image.ravel()
    image_center = tuple(s // 2 for s in image_shape)
    image_center_raveled = np.ravel_multi_index(
        image_center, image_shape
    )
    for i, offset in enumerate(offsets):
        index = image_center_raveled + offset
        assert image_raveled[index] == 0
        image_raveled[index] += i + 1

* Remove morphology._fast_pad once NumPy >= 1.17

* Add tweaks on morphology._util and testing

* Rename to _set_border_values

"edge" may be associated with "edge filters" while "border" is used
throughout the library already.

* Flatten test suite in test_util

* Manually port changes to _fast_pad from #4556

_fast_pad was changed on master in
e58cead5191939d1bc722a533c395b74c4d07528
while it was moved on this branch.

* Improve formatting and error messages