Implement SIFT feature detector and descriptor (#5472)
* added SIFT Feature Detector

This PR adds the well known SIFT feature detector and descriptor along with accompanying
demo and tests. The implementation closely follows the IPOL publication cited in the docstring
and should give very similar descriptors to the implementation accompanying that paper. The API
is implemented consistent with other feature descriptors implemented in scikit-image (e.g. providing
``detect``, ``extract`` and  ``detect_and_extract`` methods).

* refactor _find_localize_evaluate for simplicity and performance

Instead of precomputing the gradient at all points, only compute it at the positions of the maxima

* compute edge response without np.linalg.eig as in the IPOL publication

* Only compute non-redundant Hessian entries

Compute offsets as in the IPOL paper (faster to avoid use of np.linalg.solve)

* STYLE: replace use of np.hstack and np.vstack

use np.concatenate or np.stack instead for clarity about whether a new dimension is being
created.

* move hist and bins initialization outside of the loop in _compute_descriptor

* STYLE: use r, c notation for consistency with scikit-image convention

* split normalization and quantization

* Reorder axes for octaves variable within _create_scalespace

5-20% performance improvement observed

* Cythonize histogram bottleneck in _compute_descriptor

* Create a fast _local_max Cython function to replace peak_local_max calls

TST: update expected results in tests

The features are the same as before, but the order is different for the Cython implementation

* move orientation histogram bin and value computation to Cython

* increase offset_max from 0.5 to 0.6 as recommended in the IPOL paper

nominally offset of 0.5 would be shifted to a different pixel as nearest, but
Otero et. al. increased to 0.6 to avoid "possible numerical instabilities"

In practice, this change results in ~10% more keypoints for the (grayscale)
skimage.data.astronaut image.

TST: update expected outputs to reflect change in offset_max

* DOC: fix SIFT doctests and update references/notes in the docstring

* add descriptive docstrings to the Cython functions

remove some unneeded parameters from the Cython _update_histogram function signature

* Make sure _compute_descriptor uses float32 computations for float32 images

verified manually by debugging (output descriptors are always uint8 so it is hard to test directly)

ensure float32 is used internally by _compute_orientation

* Remove potentially annoying max octaves warning

describe the reduction in the docstring instead

* Update example

previous setting of trunacating keypoints can result in zero matches because matches need to
be across all keypoints.
Here we instead add a second column of figures where only a subset of the matches are displayed
for clearer visibility

* readded option to steer maximum number of octaves by self.n_octaves + renamed function to _set_number_of_octaves

* made self.deltas a property and added new member float_dtype

* improved neighbor looping in _local_max

* added description of sift

Co-authored-by: Fabian <fabian2609@aol.com>
Co-authored-by: Alexander Zeilmann <alexander.zeilmann@googlemail.com>
Co-authored-by: Gregory Lee <grlee77@gmail.com>