improve butterworth docstring and add new kwargs and gallery example (#6251)
Adding true Butterworth variant. The current Butterworth filter is actually the square of a Butterworth filter; adding a `square_butterworth` keyword argument defaulting to `True` for backwards compatibility. This argument can be set to False to allow use of the standard Butterworth. Some textbooks use the squared form, but this is not how a Butterworth is
defined in signal processing. Also added test cases verifying the expected amplitudes at the cutoff in each case, and an example using the Butterworth filter to the gallery

Co-authored-by: Alexandre de Siqueira <alex.desiqueira@igdore.org>