Add Meson-based build (#6536)
We are adding the option of Meson-based builds as an option to move away
from distutils which is scheduled for removal in Python 3.12 (PEP-632).
numpy.distutils is also deprecated.

The system is heavily adapted from SciPy's approach but unlike SciPy, we
do not need to compile Fortran code or need the BLAS/LAPACK
configuration options. To make building easier for developers, SciPy
also provides a dev.py CLI. This will likely be added next.

Using an editable install with pip ("-e" flag) is not recommended for
this build system. As this will first build with ninja and then a second
time using the old setuptools approach. So avoid unnecessary CI hours by
adapting our CI configuration accordingly.

Other noteworthy changes are some cleanup of obsolete files and 
addressing of nose-based test syntax that will be deprecated in pytest.

Co-authored-by: Jarrod Millman <jarrod.millman@gmail.com>
Co-authored-by: Matthew Brett <matthew.brett@gmail.com>
Co-authored-by: Stefan van der Walt <stefanv@berkeley.edu>
Co-authored-by: Lars Grüter <lagru@mailbox.org>