Complete deprecations targeting release 0.20 or 1.0 (#6583)
* Complete deprecation of selem parameter

in favor of footprint. Though removed_version points at 1.0, the
deprecation is finished in 0.20 which was not planned at the time of
0.19's release.

* Complete deprecation of selem module

Though removed_version points at 1.0, the deprecation is finished in
0.20 which was not planned at the time of 0.19's release.

* Complete deprecation of in_place parameter

in favor of "out".

Though removed_version points at 1.0, the deprecation is finished in
0.20 which was not planned at the time of 0.19's release.

* Complete deprecation of max_iterations parameter

in favor of "max_num_iter".

Though removed_version points at 1.0, the deprecation is finished in
0.20 which was not planned at the time of 0.19's release.

* Complete deprecation of max_iter parameter

in favor of "max_num_iter".

Though removed_version points at 1.0, the deprecation is finished in
0.20 which was not planned at the time of 0.19's release.

* Complete deprecation of iterations parameter

in favor of "num_iter".

Though removed_version points at 1.0, the deprecation is finished in
0.20 which was not planned at the time of 0.19's release.

* Complete deprecation of n_iter_max parameter

in favor of "max_num_iter".

Though removed_version points at 1.0, the deprecation is finished in
0.20 which was not planned at the time of 0.19's release.

* Complete deprecation of compute_hessian_eigenvalues

Though removed_version points at 1.0, the deprecation is finished in
0.20 which was not planned at the time of 0.19's release.

* Complete deprecation of input parameter

in favor of "label_image".

Though removed_version points at 1.0, the deprecation is finished in
0.20 which was not planned at the time of 0.19's release.

* Complete deprecation of greyco* functions

Though removed_version points at 1.0, the deprecation is finished in
0.20 which was not planned at the time of 0.19's release.

* Complete deprecation of multichannel kwarg

Though the default removed_version of deprecate_multichannel_kwarg
points at 1.0, the deprecation is finished in 0.20 which was not planned
at the time of 0.19's release.

* Remove unused deprecate_multichannel_kwarg

This decorator is no longer necessary since the deprecation of
multichannel is completed with 0.20 (previously 1.0).

Make test's independent of temporary decorations of library functions
such as hog and pyramid_gaussian. Their deprecation decorator was
removed.

* Complete deprecation of height, width in rectangle

and remove unused imports.

* Remove completed items in TODO.txt

For filters.inverse(), the removed_version targets 0.21 and not 0.20. So
move.

* Complete deprecation of neighbourhood parameter

I think removed_version="1.2" is actually supposed to be 1.0 (or 0.20).

* Remove deprecated grey and greyreconstruct modules

These deprecations are already part of 0.19. So it should be fine to
remove them in 0.20.

* Remove warning about deprecated *_iter

Missed this earlier.

* Revert "Complete deprecation of compute_hessian_eigenvalues"

This reverts commit 56f0431ea2ea2870ef7f158fcc0e0f6310c8018e.

The deprecation of `compute_hessian_eigenvalues` was introduced after
all 0.19.x releases (see git tag --contains d2689f5)! So the target
should be 0.21.

* Deprecate automatic channel detection in gaussian

Automatic detection of the color channel based on the old deprecated
`multichannel=None` was broken in version 0.19. This commit recovers
the old behavior for 0.20 using a proxy object `ChannelAxisNotSet`.
The behavior is still deprecated though and should be removed completely
in 0.21.

Note that the warning is only raised if the automatic behavior deviates
from the behavior that will be the new default in 0.21, which is
`channel_axis=None`. This way we only notify users that were affected
by the broken behavior in 0.19 and now deprecated behavior in 0.20.

* Complete deprecation of coordinates parameter

in `regionprops` (since 0.16) and `active_contour` (since 0.18).

* Remove outdated TODO for CircleModel.estimate

Instead of raising a ValueError, since
f8dfea6c3230c71ab9ddc5fde80eeb74f57a3922 [1] this is addressed by
returning `False`.

[1] https://github.com/scikit-image/scikit-image/pull/6453

* Remove removed files from meson.build

These were removed in d9517743f362aa5396fd229a06f87f33b9e3522d and
b53b46a421899eb9e0015ab4485260af300b41b7.

* Remove obsolete tests for coordinates parameter

* Catch expected warning for deprecated color channel inference in gaussian

* Update skimage/_shared/filters.py

Co-authored-by: Stefan van der Walt <sjvdwalt@gmail.com>

* Explain proxy value in more clearly

Co-authored-by: Gregory Lee <grlee77@gmail.com>
Co-authored-by: Stefan van der Walt <sjvdwalt@gmail.com>