Fix the order of return values for `hessian_matrix` (#6624)
* fix axis order in hessian_matrix to match what is documented

For mode='rc' the 2nd derivative along axis=0 should come first.
This change also makes the order consistent with that used by structure_tensor

* for consistency with structure_tensor disallow order='xy' in the 2D case

add test cases to confirm expected behavior for order

* Make 3D test of hessian_matrix more specific

Explanation: Each array in Hs is the result of

    H_elems = [np.gradient(gradients[ax0], axis=ax1)
               for ax0, ax1 in combinations_with_replacement(axes, 2)]

Because reversing axes (due to the bug fix) doesn't affect the result
of `H_elems[2]`. For index 2, `combinations_with_replacement(axes, 2)`
returns either `(0, 2)` or the reversed `(2, 0)`. Both result in the
same gradient.

So to make the test notice this change, we test another one of H_elems
arrays whose result is not the same if axes is reversed.

Co-authored-by: Lars Grüter <lagru@mailbox.org>