Move `future.graph` to `skimage.graph` (#6674)
I opted to make all moved submodules private as I expect the contained
public functions to be exposed in `skimage.graph`.

Also, I removed the duplication of `cut_normalized` in `ncut`.

Ignore "future/graph" during doctest collection.  Not doing this
will trigger the ModuleNotFoundError in
skimage/future/graph/__init__.py. The ignore_collect entry doesn't seem
to work if `graph` is a module in `skimage.future`; it works when we
turn it into a package.

It's also required to ignore "future/graph" during apigen.

Co-authored-by: Stefan van der Walt <sjvdwalt@gmail.com>